Modify API - IRBuilder::AddSymbol, IRBuilder::AddRelocation and FragmentRef::Create
1. IRBuilder::AddSymbol and IRBuilder::AddRelocation - add parameter SectionData
2. FragmentRef::Create - Replace the parameter LDSection to SectionData
3. SectionSymbolSet - add parameter TargetLDBackend

We need TargetLDBackend::isMergeStringSection to check if the section is
merge string and to acquire the MergeString from LDSection, while we do not want
to pass Backend to FragmentRef::Create. Let the caller decide the SectionData
and pass it to FragmentRef::Create directly. Hence in IRBuilder::AddSymbol and
IRBuilder::AddRelocation, we also need to pass the SectionData as its parameter
so that it's able to call FragmentRef::Create