MergeString: Refine MergeString::getOutputOffset
Add a map to MergeStringOutput to record the mapping from fragment to
MergeStringInput so that we can get the output offset of that fragment
according to the specific input offset. This fragment is recored only if it is
the first fragment in the input section, which it is the fragment that is
pointed by the input section symbol.