MergeString: Add ObjectReader::readMergeStrings and add MergeStringReader into ELFObjectReader
In some cases, there are input sections which are merged into the same output
section, but some with MergeString flags and some not. Sections with no
MergeString flags will be read in as normal sections, while in this case, we
should treat them all as merge string sections. The new API
ObjectReader::readMergeStrings helps to re-read the input sections as a
MergeString section. It will be called during MergeSections.