Relocator: add doApplyRelocation as applyRelocation hook
Relocator::applyRelocation has been split into two parts, apply relocations
for merge string sections, and apply normal relocations.

Relocator::applyRelocation is responsible to check if the relocation target
symbol is defined in the merge section or not, and invoke the corresponding
handler, doApplyRelocation or applyRelocationForMergeString. Both of
doApplyRelocation and applyRelocationForMergeString can be overwritten by
device backends, and general applyRelocationForMergeString is provided by
Relocator.