Fix ObjectLinker::relocation
When apply relocation, bypass the relocation if the target symbol
is defined in the ignore section. For example, a relocation section for
a debug section itself may not be discarded, but it's target section
would be discarded. Since we have not handle the debug information
elimination, we simply bypass the relocation.